requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-admin-inline-paginator==0.1.1
django-admin-interface==0.16.0
django-admin-rangefilter==0.7.0
django-colorfield==0.4.3
django-countries==7.2.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-modeladmin-reorder==0.3.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.3
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
phonenumbers==8.12.31
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
stripe==2.60.0
twilio==6.63.2
uritemplate==3.0.1
urllib3==1.26.6
